临时处理:上传失败情况不删除本地文件

+++ src/main/java/com/oppo/test/coverage/backend/util/file/FolderFileScanner.java
@@ -92,12 +92,17 @@ public class FolderFileScanner {
     public void fileUpload(String projectName,Long taskId){
         ArrayList<File> uploadFileList = getUploadFileList(projectName, taskId, true);
         AmazonS3 s3 = OcsUtil.getAmazonS3();
+        boolean canDelete = true;
         for (File file : uploadFileList){
             List<String> splitList = Splitter.on(".").trimResults().splitToList(file.getName());
             String extensionName = splitList.get(splitList.size()-1);
-            OcsUtil.upload(s3,file.getAbsolutePath(),file,"html".equals(extensionName)?"text/html":null);
+            if (!OcsUtil.upload(s3,file.getAbsolutePath(),file,"html".equals(extensionName)?"text/html":null)){
+                canDelete = false;
+            }
         }
-        deleteAllFileAfterUpload(projectName, taskId);
+        if (canDelete){
+            deleteAllFileAfterUpload(projectName, taskId);
+        }
     }
 
     private void deleteAllFileAfterUpload(String projectName, Long taskId){
